Move test scripts & resources to only_otters/tests
