+++ only_otters/scrapetools/hquery.py
@@ -356,25 +356,3 @@ class HierarchicalXPathQuery:
         tree = lhtml.fromstring(html)
 
         return self.__class__.process_query(tree, self.content)
-
-
-@HierarchicalXPathQuery.pipe
-def external(item):
-    print('EXTERNAL:', item)
-    return item
-
-
-@HierarchicalXPathQuery.high_pipe
-def doubidou(fn, items):
-    return ('X:' + fn(i) for i in map(str, items))
-
-
-if __name__ == "__main__":
-
-    text = open('test.html').read()
-    hxq = HierarchicalXPathQuery.from_yml('query.yml')
-
-    result = hxq(html=text)
-
-    print(*result)
-    hxq.get()
